Port tests for ArrayHasKey constraint to separate test case
