+++ tests/Framework/ConstraintTest.php
@@ -18,106 +18,6 @@ use PHPUnit\Util\Filter;
 
 class ConstraintTest extends TestCase
 {
-    public function testConstraintArrayHasKey()
-    {
-        $constraint = Assert::arrayHasKey(0);
-
-        $this->assertFalse($constraint->evaluate([], '', true));
-        $this->assertEquals('has the key 0', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate([]);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that an array has the key 0.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintArrayHasKey2()
-    {
-        $constraint = Assert::arrayHasKey(0);
-
-        try {
-            $constraint->evaluate([], 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message\nFailed asserting that an array has the key 0.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintArrayNotHasKey()
-    {
-        $constraint = Assert::logicalNot(
-            Assert::arrayHasKey(0)
-        );
-
-        $this->assertFalse($constraint->evaluate([0 => 1], '', true));
-        $this->assertEquals('does not have the key 0', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate([0 => 1]);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that an array does not have the key 0.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintArrayNotHasKey2()
-    {
-        $constraint = Assert::logicalNot(
-            Assert::arrayHasKey(0)
-        );
-
-        try {
-            $constraint->evaluate([0], 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that an array does not have the key 0.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
 
     public function testConstraintIsReadable()
     {
